chore(team): split reconcile manager tests
